EntityFlowTypeBuilder does not need special case logic in getSubFlowInfo for lava.crms.people.patient.patient and lava.crms.enrollment.status.enrollmentStatus where the add transitions to the special addPatient and addEnrollmentStatus flows respectively. This is because lava.crms.people.patient.addPatient and lava.crms.enrollment.status.addEnrollmentStatus are actions in their own right with their own set of CRUD flows, although the .add flow (event) is the only one used.
Add buttons/actions for Patient and Enrollment Status should use the "addPatient" action with "addPatient__add" event and "addEnrollmentStatus" with "addEnrollmentStatus__add" event.

NOTE: this was all fine and dandy until it broke, revealing a long-standing bug that had never surfaced. The new findPatient action has the patient action as a child but not the addPatient action, so the addPatient action is not built as part of the findPatient flow but when the findPatient flow is built a subflow transition was created to transition to the "addPatient" subflow (because that was specified in EntityFlowTypeBuilder getSubFlowInfo) and if the "addPatient" flows were not built yet, then boom, crash.