changed DEFAULT mapping keyword to STATIC
new instrExistedWithDataNoUpdate flag for each instrument to clearly
distinguish between update and no update

fixed tricky bug where an import record that will be skipped because of
an error must have all retrieved instruments evicted so they are not
persisted as dirty objects when the overall request transaction is
persisted
NOTE: may not be necessary for lava3 since transactions behave
differently (for unknown reason), as separate transactions not one
global transaction

no longer hold assumption that with a multiple instrument import the
instruments for a single data record will either all exist or all not
exist. accommodate situation where there could be a combination where
some exist with data, some exist without data, some do not exist

matchVisitType renamed to matchVisitTypeFlag as part of introducing
matchVisitType importDefinition properties to allow matching on multiple
visit types (but not using yet)

allow visitWith in the data file to just be a last name instead of last
name, first initial and attempt to match a unique LAVA user with just
last name

different technique for catching multiple instrument match which does
not result in a database error (which would mean import as a whole could
not be persisted for valid records). this is for lava2 as lava3 still
persists valid records even if one record fails on database error
because transaction handling differs for unknown reasons

do not set instrument status fields until time to save the instrument so
they will not be dirty and result in the instrument being saved before
it should be (and double save results in StaleObjectStateException)

implement support for importDefinition global default standard error
code to use for a variable if data value is blank (and there is no
variable specific default value defined in the mapping file)

implement support for variable specific default values specified in the
mapping file to assign if value in data file is blank

if instrument (with or without data) updated then update visitStatus
from SCHEDULED to COMPLETE
update instrument dc an de statuses