refactored PatientExistsHandling so birthDate validation is done after lastName and firstName validation
now handle data files with date values that incorporate a time component
into the same field value
added more validation testing that the date format used in the data file
matches the date format specified in the import definition

was not actually converting dates with 2 digit year to a date string
with 4 digit year for logging purposes

when doing a time match on visit (only when visit window is not
specified in the import definition, i.e. visit window is 0) implemented
a user settable flag as to whether or not time should be used when
attempting to match an existing visit (and this also indicates whether
the time component should be used if a new visit is created for
visitTime)

Also:
fixed bug introduced by recent change where after saving an instrument
did a refresh, invoked afterUpdate, and saved again, so that instruments
with Hibernate formulas to lookup scores would look up the score,
transfer them to persistent properties, and then save.
was working for newly created instruments, but for already existing
instruments the refresh would retrieve the pre-edited state of the
instrument, i.e. apparently saveNoFlush was not doing anything for
existing instruments.
solution was to get rid of saveNoFlush so to do that had to restructure
code so that all (instrument) updates (e.g. statuses and calculations)
are done before any saves/flushes are done. did that and voila, fixes
the problem while still doing lookups and persisting them on import